UpdateProduct path fixed

+++ src/main/java/com/example/HardwareStore/collection/Product.java
@@ -17,7 +17,7 @@ public class Product {
     @Id
     private String id;
     private String productName;
-    private Double productPrice;
+    private Integer productPrice;
     private String productDescription;
     private Integer sold;
     private Integer minUnits;
